Move the rest too
