+++ src/ProductConstructionService/ProductConstructionService.WorkItems/ProductConstructionService.WorkItems.csproj
@@ -9,7 +9,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <!-- Health check for the Azure Storage Queue fails if this is not pinned to 8.0.0. -->
+    <PackageReference Include="Aspire.Azure.Storage.Queues" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" />
     <PackageReference Include="Microsoft.Extensions.Logging" />
   </ItemGroup>
